Support creating Micro Bosh Stemcells for all stemcell types

+++ templates/ubuntu/micro.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
+set -x
 
-source _variables.sh
+bosh_app_dir=/var/vcap
+bosh_dir=${bosh_app_dir}/bosh
+infrastructure="vsphere"
+SRC_DIR=`pwd`
 
 blobstore_path=${bosh_app_dir}/micro_bosh/data/cache
 agent_host=localhost
@@ -9,15 +13,18 @@ agent_uri=http://vcap:vcap@${agent_host}:${agent_port}
 export PATH=${bosh_app_dir}/bosh/bin:$PATH
 
 # Packages
-apt-get -y install genisoimage libpq-dev
+apt-get -y install genisoimage libpq-dev nc g++ libboost-program-options-dev libboost-serialization-dev libpqclient-dev libmysqlclient-dev libsqlite3-dev
 
 # Install package compiler
-mkdir -p /tmp/package_compiler
-pushd /tmp/package_compiler
-    cp $SRC_DIR/_package_compiler.tar .
-    tar -xvf _package_compiler.tar
-    $bosh_dir/bin/gem install *.gem --no-ri --no-rdoc --local
-popd
+if [ ! -f "$bosh_dir/bin/package_compiler" ]
+then
+    mkdir -p /tmp/package_compiler
+    pushd /tmp/package_compiler
+        cp $SRC_DIR/_package_compiler.tar .
+        tar -xvf _package_compiler.tar
+        $bosh_dir/bin/gem install *.gem --no-ri --no-rdoc --local
+    popd
+fi
 
 mkdir -p ${bosh_app_dir}/bosh/blob
 mkdir -p ${blobstore_path}
